Rework segment writing to avoid race/improve perf
There was race when writing to segments under high concurrency
due to the ringbuf being fully utilized and buffers being
reused before fully flushed to disk.

To fix the race, we remove the ringbuf as straight mutexes
with buffered writer is performing better and is simpler to
reason about.

+++ ingestor/storage/store.go
@@ -150,7 +150,7 @@ func (s *LocalStore) WriteOTLPLogs(ctx context.Context, database, table string, 
 		return err
 	}
 
-	wo := wal.WithSampleMetadata(wal.LogSampleType, uint16(len(logs.Logs)))
+	wo := wal.WithSampleMetadata(wal.LogSampleType, uint32(len(logs.Logs)))
 	if err := w.Write(ctx, enc.Bytes(), wo); err != nil {
 		return err
 	}
